symfony 5 compatibility
