+++ Event/UserEvent.php
@@ -11,14 +11,13 @@
 
 namespace HWI\Bundle\OAuthBundle\Event;
 
-use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
  * @author Marek Štípek
  */
-class UserEvent extends Event
+class UserEvent extends AbstractEvent
 {
     /**
      * @var UserInterface
