+++ Tests/Fixtures/CustomOAuthToken.php
@@ -23,6 +23,6 @@ class CustomOAuthToken extends OAuthToken
             'ROLE_USER',
         ]);
 
-        $this->setUser(new FOSUser());
+        $this->setUser(class_exists(\FOS\UserBundle\Model\User::class) ? new FOSUser() : new User());
     }
 }
