+++ Tests/Form/FOSUBRegistrationFormHandlerTest.php
@@ -22,6 +22,13 @@ use Symfony\Component\HttpFoundation\Request;
 
 class FOSUBRegistrationFormHandlerTest extends TestCase
 {
+    protected function setUp(): void
+    {
+        if (!class_exists(\FOS\UserBundle\Model\User::class)) {
+            $this->markTestSkipped('FOSUserBundle not installed.');
+        }
+    }
+
     public function testProcessReturnsFalseForNotPostRequest()
     {
         $formMock = $this->getForm(false);
