+++ composer.json
@@ -94,12 +94,12 @@
 
     "require": {
         "php":                            "^7.2",
-        "symfony/framework-bundle":       "^3.4.26|^4.3",
-        "symfony/security-bundle":        "^3.4|^4.3",
-        "symfony/options-resolver":       "^3.4|^4.3",
-        "symfony/form":                   "^3.4|^4.3",
-        "symfony/yaml":                   "^3.4|^4.3",
-        "symfony/templating":             "^3.4|^4.3",
+        "symfony/framework-bundle":       "^3.4.26|^4.3|^5.0",
+        "symfony/security-bundle":        "^3.4|^4.3|^5.0",
+        "symfony/options-resolver":       "^3.4|^4.3|^5.0",
+        "symfony/form":                   "^3.4|^4.3|^5.0",
+        "symfony/yaml":                   "^3.4|^4.3|^5.0",
+        "symfony/templating":             "^3.4|^4.3|^5.0",
 
         "psr/http-message":               "^1.0",
         "php-http/client-implementation": "^1.0",
@@ -111,21 +111,21 @@
 
     "require-dev": {
         "doctrine/orm":                 "^2.3",
-        "symfony/browser-kit":          "^3.4|^4.3",
-        "symfony/css-selector":         "^3.4|^4.3",
-        "symfony/property-access":      "^3.4|^4.3",
-        "symfony/validator":            "^3.4|^4.3",
-        "symfony/twig-bundle":          "^3.4|^4.3",
-        "symfony/stopwatch":            "^3.4|^4.3",
+        "symfony/browser-kit":          "^3.4|^4.3|^5.0",
+        "symfony/css-selector":         "^3.4|^4.3|^5.0",
+        "symfony/property-access":      "^3.4|^4.3|^5.0",
+        "symfony/validator":            "^3.4|^4.3|^5.0",
+        "symfony/twig-bundle":          "^3.4|^4.3|^5.0",
+        "symfony/stopwatch":            "^3.4|^4.3|^5.0",
         "symfony/phpunit-bridge":       "^5.0",
-        "symfony/translation":          "^3.4|^4.3",
+        "symfony/translation":          "^3.4|^4.3|^5.0",
         "friendsofsymfony/user-bundle": "^2.1",
         "php-http/httplug-bundle":      "^1.7",
         "php-http/guzzle6-adapter":     "^2.0",
         "phpunit/phpunit":              "^6.5|^7.5",
         "friendsofphp/php-cs-fixer":    "^2.0",
         "symfony/monolog-bundle":       "^3.4",
-        "doctrine/doctrine-bundle":     "^1.11"
+        "doctrine/doctrine-bundle":     "^1.11|^2.0"
     },
 
     "conflict": {
